move fees tests to test module
